# api/openapi.yaml
openapi: 3.1.0
info:
  title: Yoon API
  version: 0.1.0
  summary: One API in front of African payment providers.
  description: |
    Yoon is a self-hosted payment gateway. Your application sends **intent** — collect
    5 000 XOF from this customer by Wave — and Yoon picks a configured provider, talks to
    it, verifies its callbacks and keeps a ledger.

    ## Conventions
    - **Authentication:** `Authorization: Bearer yk_…` on every `/v1` route except `/v1/about`.
    - **Money:** integer amounts in the currency's minor unit. XOF has none: 5 000 XOF is `5000`.
      Never floating point.
    - **Idempotency:** every `POST` requires an `Idempotency-Key` header. Retrying with the same
      key and body returns the original response (with `Idempotent-Replayed: true`). The same key
      with a different body is refused. Keys are kept for 24 hours.
    - **Errors:** [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457) `application/problem+json`
      with a stable machine-readable `code`.
    - **Lists:** newest first, cursor-paginated with `starting_after` and `limit` (max 100).
    - **Phone numbers** are returned masked (`+22177***67`).
    - **Unknown outcomes:** if a provider does not answer, the payment stays `pending` on that
      provider and is never sent to another one — the customer may already be paying.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
security:
  - apiKey: []
...
paths:
...
  /checkout/api/{id}:
    get:
      tags: [Hosted checkout]
      operationId: getCheckout
      summary: The checkout page's view of a payment (public, token-scoped)
      description: |
        What the customer's browser may know about a hosted-checkout payment: amount, description,
        status, the methods it may pick, and what to do next. Nothing else. Cheap enough to poll.
        An unknown id, a missing token and a wrong token are the same 404.
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/CheckoutToken'
      responses:
        '200':
          description: The checkout.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutView' }
        '404': { $ref: '#/components/responses/CheckoutNotFound' }
        '429': { $ref: '#/components/responses/CheckoutRateLimited' }
  /checkout/api/{id}/attempts:
    post:
      tags: [Hosted checkout]
      operationId: chooseCheckoutMethod
      summary: The customer picks a method (public, token-scoped)
      description: |
        Starts one provider attempt round for the chosen method through Yoon's routing and
        failover. Only while the payment is `created` and no other round is running: a second
        click gets 409 `checkout_in_progress`, a payment already with a provider (including an
        unknown outcome) gets 409 `checkout_not_open`. If every provider refuses, the payment stays
        `created` and `last_error` says why; the customer may choose again.
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/CheckoutToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutAttemptRequest' }
            example: { method: wave }
      responses:
        '200':
          description: The round ended; the checkout as it is now.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/CheckoutNotFound' }
        '409':
          description: |
            `checkout_in_progress` (another attempt is running), `checkout_not_open` (the payment is
            no longer waiting for a choice) or `checkout_expired`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: '`method_unavailable`: this method is not offered for this payment right now.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/CheckoutRateLimited' }
  /admin/v1/applications:
    get:
      tags: [Admin]
      operationId: adminListApplications
      summary: Applications on this instance
      security: [{ adminToken: [] }]
      responses:
        '200':
          description: Every application, by name.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required: [id, object, name, created_at]
                      properties:
                        id: { type: string, format: uuid }
                        object: { type: string, const: application }
                        name: { type: string }
                        created_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/payments:
    get:
      tags: [Admin]
      operationId: adminListPayments
      summary: One application's payments
      description: Read-only; the same results the application gets from `GET /v1/payments`.
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PaymentStatus' } }
        - { name: reference, in: query, schema: { type: string } }
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of payments, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/payments/{id}/events:
    get:
      tags: [Admin]
      operationId: adminListPaymentEvents
      summary: Status history of one payment
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { $ref: '#/components/responses/StatusHistory' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/refunds:
    get:
      tags: [Admin]
      operationId: adminListRefunds
      summary: One application's refunds
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/RefundStatus' } }
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of refunds, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RefundPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/refunds/{id}/events:
    get:
      tags: [Admin]
      operationId: adminListRefundEvents
      summary: Status history of one refund
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { $ref: '#/components/responses/StatusHistory' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/payouts:
    get:
      tags: [Admin]
      operationId: adminListApplicationPayouts
      summary: One application's payouts
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PayoutStatus' } }
        - { name: reference, in: query, schema: { type: string } }
        - { name: needs_review, in: query, schema: { type: boolean } }
        - $ref: '#/components/parameters/StartingAfter'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of payouts, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayoutPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/payouts/{id}/events:
    get:
      tags: [Admin]
      operationId: adminListPayoutEvents
      summary: Status history of one payout
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { $ref: '#/components/responses/StatusHistory' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/balances:
    get:
      tags: [Admin]
      operationId: adminGetBalances
      summary: One application's balances per provider account
      description: Yoon's view from its own ledger, not the provider's statement.
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Balances.
          content:
            application/json:
              schema:
                type: object
                required: [object, note, data]
                properties:
                  object: { type: string, const: balances }
                  note: { type: string }
                  data:
                    type: array
                    items:
                      type: object
                      required: [account, provider, currency, amount]
                      properties:
                        account: { type: string }
                        provider: { type: string }
                        currency: { type: string }
                        amount: { type: integer, format: int64 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /admin/v1/applications/{application_id}/ledger/entries:
    get:
      tags: [Admin]
      operationId: adminListLedgerEntries
      summary: One application's ledger entries
      security: [{ adminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - { name: provider, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: starting_after, in: query, schema: { type: integer, format: int64 }, description: Cursor from `next_cursor`. }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of entries, newest first. Positive = debit, negative = credit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LedgerEntryPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: 'An application API key: `yk_…`. Create one with `yoon apps create <name>`.'
    adminToken:
      type: http
      scheme: bearer
      description: The operator token set in `YOON_ADMIN_TOKEN`.
...
  parameters:
    ApplicationId:
      name: application_id
      in: path
      required: true
      description: From `GET /admin/v1/applications`.
      schema: { type: string, format: uuid }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    CheckoutToken:
      name: Yoon-Checkout-Token
      in: header
      required: true
      description: The `t` value of the checkout page URL. Missing or wrong → 404.
      schema: { type: string, maxLength: 128 }
    StartingAfter:
      name: starting_after
      in: query
      description: Cursor — the `next_cursor` of the previous page.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: Invalid request (`invalid_request`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid API key (`unauthorized`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    CheckoutNotFound:
      description: Unknown payment, not a hosted checkout, or missing/wrong token — indistinguishable (`resource_not_found`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    CheckoutRateLimited:
      description: Too many checkout requests from this address (`rate_limited`); retry after `Retry-After` seconds.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    StatusHistory:
      description: Every status change and ignored late news, oldest first.
      content:
        application/json:
          schema:
            type: array
            items: { $ref: '#/components/schemas/StatusEvent' }
    AdminDisabled:
      description: The admin API is disabled (no `YOON_ADMIN_TOKEN`), or the resource does not exist.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable, machine-readable error code.
    PaymentPage:
      type: object
      required: [data, has_more, next_cursor]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Payment' } }
        has_more: { type: boolean }
        next_cursor: { $ref: '#/components/schemas/NextCursor' }
    RefundPage:
      type: object
      required: [data, has_more, next_cursor]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Refund' } }
        has_more: { type: boolean }
        next_cursor: { $ref: '#/components/schemas/NextCursor' }
    PayoutPage:
      type: object
      required: [data, has_more, next_cursor]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Payout' } }
        has_more: { type: boolean }
        next_cursor: { $ref: '#/components/schemas/NextCursor' }
    LedgerEntryPage:
      type: object
      required: [data, has_more, next_cursor]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/LedgerEntry' } }
        has_more: { type: boolean }
        next_cursor: { $ref: '#/components/schemas/NextCursor' }
...
    NextCursor:
      type: [string, 'null']
      description: Pass as `starting_after` to get the next page; null on the last page.
    LedgerEntry:
      type: object
      required: [id, posting_id, description, account, currency, amount, created_at]
      description: Positive = debit, negative = credit.
      properties:
        id: { type: integer, format: int64 }
        posting_id: { type: integer, format: int64 }
        description: { type: string }
        account: { type: string, example: 'provider_balance:paydunya' }
        currency: { type: string }
        amount: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
    PaymentStatus:
      type: string
      enum: [created, pending, succeeded, failed, expired]
      description: |
        `pending` — with the provider, waiting for the customer (or for the provider's answer).
        `succeeded` is final. `failed` and `expired` can still become `succeeded` if the provider
        later confirms the money arrived.
    RefundStatus:
      type: string
      enum: [created, pending, unknown, refunded, failed]
    PayoutStatus:
      type: string
      enum: [created, processing, unknown, paid, failed]
    Failure:
      type: [object, 'null']
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    Payment:
      type: object
      required: [id, object, status, amount, amount_refunded, currency, country, method, created_at, updated_at]
      properties:
        id: { type: string, example: pay_0199a3f0c2e47a1b9c3d5e6f7a8b9c0d }
        object: { type: string, const: payment }
        status: { $ref: '#/components/schemas/PaymentStatus' }
        amount: { type: integer, format: int64 }
        amount_refunded: { type: integer, format: int64 }
        currency: { type: string }
        country: { type: string }
        method:
          type: string
          example: wave
          description: >-
            The payment method. `any` for a hosted checkout whose customer has not chosen yet
            (kept a string so clients built against 0.1.0 keep working).
        checkout:
          type: string
          enum: [direct, hosted]
          description: '`hosted`: the customer picks the method on the Yoon page at `checkout_url`.'
        checkout_expires_at:
          type: [string, 'null']
          format: date-time
          description: Hosted checkout only. After this an unused checkout fails with `checkout_expired`.
        reference: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        customer:
          type: [object, 'null']
          properties:
            phone: { type: [string, 'null'], example: '+22177***67' }
        provider: { type: [string, 'null'] }
        provider_reference: { type: [string, 'null'] }
        checkout_url:
          type: [string, 'null']
          description: |
            Redirect the customer here, if set. For a hosted checkout, always the Yoon page
            (`{YOON_PUBLIC_URL}/checkout/{id}?t=…`); it forwards the customer to the provider.
        instructions: { type: [string, 'null'], description: Push/USSD instruction to show the customer, if set. }
        routing_reason: { type: [string, 'null'], description: Why this provider — and any rejections before it. }
        failure: { $ref: '#/components/schemas/Failure' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CheckoutView:
      type: object
      description: What the hosted checkout page shows. Holds no secret and no other object.
      required: [id, status, amount, currency, country, description, method, methods, can_choose,
                 next_action, customer_phone, has_pi_alias, return_url, expires_at, last_error]
      properties:
        id: { type: string }
        status: { $ref: '#/components/schemas/PaymentStatus' }
        amount: { type: integer, format: int64, description: Minor units. }
        currency: { type: string }
        country: { type: string }
        description: { type: [string, 'null'] }
        method: { type: [string, 'null'], description: The method chosen last. }
        methods:
          type: array
          items:
            type: object
            required: [method, available, needs]
            properties:
              method: { type: string, example: wave }
              available: { type: boolean, description: False while every provider for it is unavailable. }
              needs:
                type: string
                enum: [none, pi_alias]
                description: '`pi_alias`: ask the customer for their PI-SPI alias (unless `has_pi_alias`).'
        can_choose: { type: boolean, description: True while the customer may start an attempt. }
        next_action:
          type: [object, 'null']
          required: [type, url, instructions]
          properties:
            type: { type: string, enum: [redirect, instructions] }
            url: { type: [string, 'null'], description: The provider's checkout page. }
            instructions: { type: [string, 'null'], description: Push/USSD instructions from the provider. }
        customer_phone: { type: [string, 'null'], example: '+22177***67', description: Masked. }
        has_pi_alias: { type: boolean }
        return_url: { type: [string, 'null'], description: "The application's page to go back to." }
        expires_at: { type: [string, 'null'], format: date-time }
        last_error:
          type: [object, 'null']
          description: Why the last round was refused by every provider (only right after it).
          required: [code]
          properties:
            code: { type: string, example: PHONE_REQUIRED }
    CheckoutAttemptRequest:
      type: object
      required: [method]
      properties:
        method: { type: string, maxLength: 32 }
        phone: { type: string, maxLength: 32, description: 'Optional; validated for the payment country.' }
        pi_alias: { type: string, maxLength: 64 }
...
    Refund:
      type: object
      required: [id, object, payment_id, status, amount, currency, provider, created_at, updated_at]
      properties:
        id: { type: string, example: re_0199a3f0c2e47a1b9c3d5e6f7a8b9c0d }
        object: { type: string, const: refund }
        payment_id: { type: string }
        status: { $ref: '#/components/schemas/RefundStatus' }
        amount: { type: integer, format: int64 }
        currency: { type: string }
        reason: { type: [string, 'null'] }
        provider: { type: string }
        provider_reference: { type: [string, 'null'] }
        failure: { $ref: '#/components/schemas/Failure' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Payout:
      type: object
      required: [id, object, status, amount, currency, country, method, recipient, provider, needs_review, created_at, updated_at]
      properties:
        id: { type: string, example: po_0199a3f0c2e47a1b9c3d5e6f7a8b9c0d }
        object: { type: string, const: payout }
        status: { $ref: '#/components/schemas/PayoutStatus' }
        amount: { type: integer, format: int64 }
        currency: { type: string }
        country: { type: string }
        method: { type: string }
        reference: { type: [string, 'null'] }
        recipient:
          type: object
          required: [phone]
          properties:
            phone: { type: [string, 'null'], example: '+22177***67', description: Masked. Null when the payout went to a PI-SPI alias. }
        provider: { type: string }
        provider_reference: { type: [string, 'null'] }
        routing_reason: { type: [string, 'null'] }
        needs_review: { type: boolean, description: The outcome stayed unknown too long; a human should check with the provider. }
        failure: { $ref: '#/components/schemas/Failure' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    StatusEvent:
      type: object
      required: [to_status, decision, cause, created_at]
      properties:
        from_status: { type: [string, 'null'] }
        to_status: { type: string }
        decision:
          type: string
          enum: [apply, ignore]
          description: '`ignore` — recorded but not applied (a duplicate, or late news such as a failure after success).'
        cause: { type: string, enum: [api, provider_call, webhook, sweep, admin] }
        raw_status: { type: [string, 'null'], description: The provider's own status value. }
        detail: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }